Restructure files to make bot.py more uncluttered
