+++ bot.py
@@ -1,10 +1,13 @@
-import aiohttp, discord, asyncio, json, yaml, logging, sys, imgkit, os, pytz, traceback
-from datetime import datetime
-from discord.ext.commands.core import command
-from itertools import islice
+import aiohttp, discord, asyncio, json, yaml, logging, sys
 from discord import Webhook, AsyncWebhookAdapter
-from discord.ext import commands, tasks
-from infoscraper import streamInfo, channelInfo
+from discord.ext import commands
+from ext.infoscraper import channelInfo
+from ext.cogs.subCycle import StreamCycle, streamcheck
+from ext.cogs.msCycle import msCycle, milestoneNotify
+from ext.share.botUtils import *
+from ext.share.dataGrab import *
+from ext.share.dataWrite import *
+from ext.share.prompts import *
 
 with open("data/settings.yaml") as f:
     settings = yaml.load(f, Loader=yaml.FullLoader)
@@ -16,451 +19,6 @@ elif settings["logging"] == "debug":
 
 bot = commands.Bot(command_prefix=commands.when_mentioned_or(settings["prefix"]), help_command=None)
 bot.remove_command('help')
-
-def chunks(data, SIZE=10000):
-    it = iter(data)
-    for i in range(0, len(data), SIZE):
-        yield {k:data[k] for k in islice(it, SIZE)}
-
-def creatorCheck(ctx):
-    return ctx.author.id == 256009740239241216
-
-def subPerms(ctx):
-    userPerms = ctx.channel.permissions_for(ctx.author)
-    return userPerms.administrator or userPerms.manage_webhooks or ctx.guild.owner_id == ctx.author.id
-
-async def subCheck(ctx, subMsg, mode, chName):
-    if mode == 1:
-        action = "Subscribe"
-    elif mode == 2:
-        action = "Unsubscribe"
-    else:
-        return {
-            "success": False
-        }
-    
-    subEmbed = discord.Embed(title=chName, description=f"{action} to the channel's:\n\n"
-                                                        "1. Livestream Notifications\n2. Milestone Notifications\n3. Both\n\n"
-                                                        "X. Cancel\n\n[Bypass this by setting the channel's default subscription type using `y!subdefault`]")
-    
-    if mode == 2:
-        subEmbed.description=f"Unsubscribe to channel with subscription type:\n\n" \
-                              "1. Livestream Notifications\n2. Milestone Notifications\n3. Both\n\n" \
-                              "X. Cancel\n\n[Bypass this by setting the channel's default subscription type using `y!subdefault`]"
-
-    await subMsg.edit(content=None, embed=subEmbed)
-
-    uInput = {
-        "success": False
-    }
-
-    def check(m):
-        return m.content.lower() in ['1', '2', '3', 'x'] and m.author == ctx.author
-
-    while True:
-        try:
-            msg = await bot.wait_for('message', timeout=60.0, check=check)
-        except asyncio.TimeoutError:
-            await subMsg.delete()
-            break
-        else:
-            if msg.content in ['1', '2', '3']:
-                await msg.delete()
-                if msg.content == '1':
-                    uInput["subType"] = ["livestream"]
-                elif msg.content == '2':
-                    uInput["subType"] = ["milestone"]
-                elif msg.content == '3':
-                    uInput["subType"] = ["livestream", "milestone"]
-                uInput["success"] = True
-                break
-            elif msg.content.lower() == 'x':
-                await msg.delete()
-                break
-            else:
-                await msg.delete()
-    
-    return uInput
-
-async def getSubType(ctx, mode, prompt = None):
-    pEmbed = discord.Embed()
-
-    if mode == 1:
-        pEmbed.title = "Default Channel Subscription Type"
-        pEmbed.description = "Set this to:\n\n1. Livestream Notifications\n2. Milestone Notifications\n" \
-                             "3. Both\n\nX. Cancel"
-        def check(m):
-            return m.content.lower() in ['1', '2', '3', 'x'] and m.author == ctx.author
-        prompt = await ctx.send(embed=pEmbed)
-    elif mode == 2:
-        pEmbed.title = "Channel Subscription Type"
-        pEmbed.description = "Get subscription list for:\n\n1. Livestream Notifications\n" \
-                             "2. Milestone Notifications\n\nX. Cancel" \
-                             "\n\n[Bypass this by setting the channel's default subscription type using `y!subdefault`]"
-        def check(m):
-            return m.content.lower() in ['1', '2', 'x'] and m.author == ctx.author  
-        await prompt.edit(content=None, embed=pEmbed)
-
-    while True:
-        try:
-            msg = await bot.wait_for('message', timeout=60.0, check=check)
-        except asyncio.TimeoutError:
-            await prompt.delete()
-            await ctx.message.delete()
-            return {
-                "success": False
-            }
-        else:
-            if msg.content in ['1', '2', '3']:
-                subTypes = ["livestream", "milestone"]
-                if mode == 1:
-                    with open("data/servers.json") as f:
-                        servers = json.load(f)
-                    if msg.content != '3':
-                        subType = [subTypes[int(msg.content) - 1]]
-                        subText = subTypes[int(msg.content) - 1]
-                    else:
-                        subType = subTypes
-                        subText = "both"
-                    try:
-                        servers[str(ctx.guild.id)][str(ctx.channel.id)]["subDefault"] = subType
-                    except:
-                        servers = await genServer(servers, ctx.guild, ctx.channel)
-                        servers[str(ctx.guild.id)][str(ctx.channel.id)]["subDefault"] = subType
-                    with open("data/servers.json", "w") as f:
-                        json.dump(servers, f, indent=4)
-
-                    await msg.delete()
-                    await ctx.message.delete()
-                    await prompt.edit(content=f"This channel will now subscribe to {subText} notifications by default.", embed=None)
-                    break
-                elif mode == 2 and msg.content in ['1', '2']:
-                    await msg.delete()
-                    return {
-                        "success": True,
-                        "subType": subTypes[int(msg.content) - 1]
-                    }
-                else:
-                    await msg.delete()
-            elif msg.content.lower() == 'x':
-                await msg.delete()
-                await prompt.delete()
-                await ctx.message.delete()
-                return {
-                    "success": False
-                }
-            else:
-                await msg.delete()
-
-async def botError(ctx, error):
-    errEmbed = discord.Embed(title="An error has occurred!", color=discord.Colour.red())
-    if "403 Forbidden" in str(error):
-        permData = [{
-            "formatName": "Manage Webhooks",
-            "dataName": "manage_webhooks"
-        }, {
-            "formatName": "Manage Messages",
-            "dataName": "manage_messages"
-        }]
-        permOutput = []
-        for perm in iter(ctx.guild.me.permissions_in(ctx.channel)):
-            for pCheck in permData:
-                if perm[0] == pCheck["dataName"]:
-                    if not perm[1]:
-                        permOutput.append(pCheck["formatName"])
-        plural = "this permission"
-        if len(permOutput) > 1:
-            plural = "these permissions"
-        errEmbed.description = "This bot has insufficient permissions for this channel.\n" \
-                               f"Please allow the bot {plural}:\n"
-        for perm in permOutput:
-            errEmbed.description += f'\n - `{perm}`'
-        
-        return errEmbed
-    elif isinstance(error, commands.CheckFailure):
-        errEmbed.description = "You are missing permissions to use this bot.\n" \
-                               "Ensure that you have one of these permissions for the channel/server:\n\n" \
-                               " - `Administrator (Server)`\n - `Manage Webhooks (Channel/Server)`"
-        
-        return errEmbed
-    else:
-        print("An unknown error has occurred.")
-        traceback.print_tb(error.__traceback__)
-        print(error)
-
-async def genServer(servers, cserver, cchannel):
-    if str(cserver.id) not in servers:
-        logging.debug("New server! Adding to database.")
-        servers[str(cserver.id)] = {
-            str(cchannel.id): {
-                "url": "",
-                "notified": {},
-                "livestream": [],
-                "milestone": []
-            }
-        }
-    elif str(cchannel.id) not in servers[str(cserver.id)]:
-        logging.debug("New channel in server! Adding to database.")
-        servers[str(cserver.id)][str(cchannel.id)] = {
-            "url": "",
-            "notified": {},
-            "livestream": [],
-            "milestone": []
-        }
-    
-    return servers
-
-async def getwebhook(servers, cserver, cchannel):
-    if isinstance(cserver, str) and isinstance(cchannel, str):
-        cserver = bot.get_guild(int(cserver))
-        cchannel = bot.get_channel(int(cchannel))
-    try:
-        logging.debug(f"Trying to get webhook url for {cserver.id}")
-        whurl = servers[str(cserver.id)][str(cchannel.id)]["url"]
-    except KeyError:
-        logging.debug("Failed to get webhook url! Creating new webhook.")
-        servers = await genServer(servers, cserver, cchannel)
-        with open("yagoo.jpg", "rb") as image:
-            webhook = await cchannel.create_webhook(name="Yagoo", avatar=image.read())
-        whurl = webhook.url
-        servers[str(cserver.id)][str(cchannel.id)]["url"] = whurl
-        with open("data/servers.json", "w") as f:
-            json.dump(servers, f, indent=4)
-    return whurl
-
-async def streamcheck(ctx = None, test: bool = False, loop: bool = False):
-    with open("data/channels.json", encoding="utf-8") as f:
-        channels = json.load(f)
-    if not test:
-        cstreams = {}
-        for channel in channels:
-            for x in range(2):
-                try:
-                    # print(f'Checking {cshrt["name"]}...')
-                    if channel != "":
-                        status = await streamInfo(channel)
-                        ytchannel = await channelInfo(channel)
-                        if status["isLive"]:
-                            cstreams[channel] = {
-                                "name": ytchannel["name"],
-                                "image": ytchannel["image"],
-                                "videoId": status["videoId"],
-                                "videoTitle": status["videoTitle"],
-                                "timeText": status["timeText"]
-                            }
-                    break
-                except:
-                    continue
-        return cstreams
-    else:
-        stext = ""
-        stext2 = ""
-        for channel in channels:
-            for x in range(2):
-                try:
-                    # print(f'Checking {cshrt["name"]}...')
-                    if channel != "":
-                        status = await streamInfo(channel)
-                        ytchan = await channelInfo(channel)
-                        if len(stext) + len(f'{ytchan["name"]}: <:green_circle:786380003306111018>\n') <= 2000:
-                            if status["isLive"]:
-                                stext += f'{ytchan["name"]}: <:green_circle:786380003306111018>\n'
-                            else:
-                                stext += f'{ytchan["name"]}: <:red_circle:786380003306111018>\n'
-                        else:
-                            if status["isLive"]:
-                                stext2 += f'{ytchan["name"]}: <:green_circle:786380003306111018>\n'
-                            else:
-                                stext2 += f'{ytchan["name"]}: <:red_circle:786380003306111018>\n'
-                    break
-                except:
-                    if x == 2:
-                        if len(stext) + len(f'{channel}: <:warning:786380003306111018>\n') <= 2000:
-                            stext += f'{channel}: <:warning:786380003306111018>\n'
-                        else:
-                            stext2 += f'{channel}: <:warning:786380003306111018>\n'
-        await ctx.send(stext.strip())
-        await ctx.send(stext2.strip())
-
-async def streamNotify(cData):
-    with open("data/servers.json", encoding="utf-8") as f:
-        servers = json.load(f)
-    for server in servers:
-        for channel in servers[server]:
-            for ytch in cData:
-                if ytch not in servers[server][channel]["notified"] and ytch in servers[server][channel]["livestream"]:
-                    servers[server][channel]["notified"][ytch] = {
-                        "videoId": ""
-                    }
-                if ytch in servers[server][channel]["livestream"] and cData[ytch]["videoId"] != servers[server][channel]["notified"][ytch]["videoId"]:
-                    whurl = await getwebhook(servers, server, channel)
-                    async with aiohttp.ClientSession() as session:
-                        embed = discord.Embed(title=f'{cData[ytch]["videoTitle"]}', url=f'https://youtube.com/watch?v={cData[ytch]["videoId"]}')
-                        embed.description = f'Started streaming {cData[ytch]["timeText"]}'
-                        embed.set_image(url=f'https://img.youtube.com/vi/{cData[ytch]["videoId"]}/maxresdefault.jpg')
-                        webhook = Webhook.from_url(whurl, adapter=AsyncWebhookAdapter(session))
-                        await webhook.send(f'New livestream from {cData[ytch]["name"]}!', embed=embed, username=cData[ytch]["name"], avatar_url=cData[ytch]["image"])
-                        servers[server][channel]["notified"][ytch]["videoId"] = cData[ytch]["videoId"]
-    with open("data/servers.json", "w", encoding="utf-8") as f:
-        servers = json.dump(servers, f, indent=4)
-
-async def streamClean(cData):
-    with open("data/servers.json", encoding="utf-8") as f:
-        servers = json.load(f)
-    livech = []
-    for ytch in cData:
-        livech.append(ytch)
-    for server in servers:
-        for channel in servers[server]:
-            for ytch in servers[server][channel]["notified"]:
-                if ytch not in livech:
-                    servers[server][channel]["notified"].remove(ytch)
-    with open("data/servers.json", "w", encoding="utf-8") as f:
-        servers = json.dump(servers, f, indent=4)
-    
-class StreamCycle(commands.Cog):
-    def __init__(self):
-        self.timecheck.start()
-
-    def cog_unload(self):
-        self.timecheck.cancel()
-
-    @tasks.loop(minutes=3.0)
-    async def timecheck(self):
-        logging.info("Starting stream checks.")
-        cData = await streamcheck(loop=True)
-        logging.info("Notifying channels (Stream).")
-        await streamNotify(cData)
-        logging.info("Stream checks done.")
-
-async def milestoneCheck():
-    with open("data/channels.json") as f:
-        channels = json.load(f)
-    
-    milestone = {}
-    noWrite = True
-
-    for channel in channels:
-        for x in range(2):
-            try:
-                ytch = await channelInfo(channel)
-                logging.debug(f'Milestone - Checking channel: {ytch["name"]}')
-                if ytch["roundSubs"] > channels[channel]["milestone"]:
-                    noWrite = False
-                    if ytch["roundSubs"] < 1000000:
-                        subtext = f'{int(ytch["roundSubs"] / 1000)}K Subscribers'
-                    else:
-                        if ytch["roundSubs"] == ytch["roundSubs"] - (ytch["roundSubs"] % 1000000):
-                            subtext = f'{int(ytch["roundSubs"] / 1000000)}M Subscribers'
-                        else:
-                            subtext = f'{ytch["roundSubs"] / 1000000}M Subscribers'
-                    milestone[channel] = {
-                        "name": ytch["name"],
-                        "image": ytch["image"],
-                        "banner": ytch["mbanner"],
-                        "msText": subtext
-                    }
-                    channels[channel]["milestone"] = ytch["roundSubs"]
-                    break
-            except:
-                if x == 2:
-                    logging.error(f'Milestone - Unable to get info for {channel}!')
-                    break
-                else:
-                    logging.warning(f'Milestone - Failed to get info for {channel}. Retrying...')
-    
-    if not noWrite:
-        with open("data/channels.json", "w") as f:
-            json.dump(channels, f, indent=4)
-    
-    return milestone
-
-async def milestoneNotify(msDict):
-    logging.debug(f'Milestone Data: {msDict}')
-    with open("data/servers.json") as f:
-        servers = json.load(f)
-    for channel in msDict:
-        logging.debug(f'Generating milestone image for id {channel}')
-        with open("milestone/milestone.html") as f:
-            msHTML = f.read()
-        options = {
-            "enable-local-file-access": "",
-            "encoding": "UTF-8",
-            "quiet": ""
-        }
-        msHTML = msHTML.replace('[msBanner]', msDict[channel]["banner"]).replace('[msImage]', msDict[channel]["image"]).replace('[msName]', msDict[channel]["name"]).replace('[msSubs]', msDict[channel]["msText"])
-        logging.debug(f'Replaced HTML code')
-        with open("milestone/msTemp.html", "w", encoding="utf-8") as f:
-            f.write(msHTML)
-        logging.debug(f'Generating image for milestone')
-        if not os.path.exists("milestone/generated"):
-            os.mkdir("milestone/generated")
-        imgkit.from_file("milestone/msTemp.html", f'milestone/generated/{channel}.png', options=options)
-        logging.debug(f'Removed temporary HTML file')
-        os.remove("milestone/msTemp.html")
-        for server in servers:
-            logging.debug(f'Accessing server id {server}')
-            for dch in servers[server]:
-                logging.debug(f'Milestone - Channel Data: {servers[server][dch]["milestone"]}')
-                logging.debug(f'Milestone - Channel Check Pass: {channel in servers[server][dch]["milestone"]}')
-                if channel in servers[server][dch]["milestone"]:
-                    logging.debug(f'Posting to {dch}...')
-                    await bot.get_channel(int(dch)).send(f'{msDict[channel]["name"]} has reached {msDict[channel]["msText"].replace("Subscribers", "subscribers")}!', file=discord.File(f'milestone/generated/{channel}.png'))
-                    await bot.get_channel(int(dch)).send("おめでとう！")
-
-class msCycle(commands.Cog):
-    def __init__(self):
-        self.timecheck.start()
-
-    def cog_unload(self):
-        self.timecheck.cancel()
-
-    @tasks.loop(minutes=3.0)
-    async def timecheck(self):
-        logging.info("Starting milestone checks.")
-        msData = await milestoneCheck()
-        if msData != {}:
-            logging.info("Notifying channels (Milestone).")
-            await milestoneNotify(msData)
-        logging.info("Milestone checks done.")
-
-async def bdayCheck():
-    update = False
-    write = False
-    bdayResults = {}
-
-    with open("data/bot.json") as f:
-        bdata = json.load(f)
-    
-    if "bdayCheck" not in bdata:
-        bdata["bdayCheck"] = datetime.now(pytz.timezone("Asia/Tokyo")).strftime("%m%d")
-        write = True
-
-    now = datetime.now(pytz.timezone("Asia/Tokyo")).replace(hour=0, minute=0, second=0, microsecond=0)
-    if pytz.timezone("Asia/Tokyo").localize(datetime.strptime(bdata["bdayCheck"], "%m%d").replace(year=datetime.now(pytz.timezone("Asia/Tokyo")).year)) < now:
-        with open("data/birthdays.json") as f:
-            bdays = json.load(f)
-
-        if str(now.day) in bdays[str(now.month)]:
-            bdayCurrent = bdays[str(now.month)][str(now.day)]
-            update = True
-        
-        bdata["bdayCheck"] = datetime.now(pytz.timezone("Asia/Tokyo")).strftime("%m%d")
-        write = True
-    
-    if write:
-        with open("data/bot.json") as f:
-            json.dump(bdata, f)
-    
-    if update:
-        return {
-            "occurToday": update,
-            "bdayData": bdayCurrent
-        }
-    else:
-        return {
-            "occurToday": update
-        }
 
 @bot.event
 async def on_ready():
@@ -503,7 +61,7 @@ async def help(ctx):
 @bot.command(aliases=['subdefault'])
 @commands.check(subPerms)
 async def subDefault(ctx):
-    await getSubType(ctx, 1)
+    await getSubType(ctx, 1, bot)
 
 @subDefault.error
 async def subdef_error(ctx, error):
@@ -511,7 +69,6 @@ async def subdef_error(ctx, error):
     if errEmbed:
         await ctx.send(embed=errEmbed)
 
-# TODO: Error on missing perms (Manage Webhooks and Manage Messages) here
 @bot.command(aliases=['sub'])
 @commands.check(subPerms)
 async def subscribe(ctx):
@@ -578,12 +135,12 @@ async def subscribe(ctx):
                 if msg.content in ['1', '2', '3', '4', '5', '6', '7', '8', '9']:
                     with open("data/servers.json") as f:
                         servers = json.load(f)
-                    await getwebhook(servers, ctx.guild, ctx.channel)
+                    await getwebhook(bot, servers, ctx.guild, ctx.channel)
                     with open("data/servers.json") as f:
                         servers = json.load(f)
                     await msg.delete()
                     if "subDefault" not in servers[str(ctx.guild.id)][str(ctx.channel.id)]:
-                        uInput = await subCheck(ctx, listmsg, 1, csplit[pagepos][picklist[int(msg.content) - 1]]["name"])
+                        uInput = await subCheck(ctx, bot, listmsg, 1, csplit[pagepos][picklist[int(msg.content) - 1]]["name"])
                     else:
                         uInput = {
                             "success": True,
@@ -616,12 +173,12 @@ async def subscribe(ctx):
                 elif msg.content.lower() == 'a':
                     with open("data/servers.json") as f:
                         servers = json.load(f)
-                    await getwebhook(servers, ctx.guild, ctx.channel)
+                    await getwebhook(bot, servers, ctx.guild, ctx.channel)
                     with open("data/servers.json") as f:
                         servers = json.load(f)
                     await msg.delete()
                     if "subDefault" not in servers[str(ctx.guild.id)][str(ctx.channel.id)]:
-                        uInput = await subCheck(ctx, listmsg, 1, "Subscribing to all channels.")
+                        uInput = await subCheck(ctx, bot, listmsg, 1, "Subscribing to all channels.")
                     else:
                         uInput = {
                             "success": True,
@@ -662,7 +219,6 @@ async def sub_error(ctx, error):
     if errEmbed:
         await ctx.send(embed=errEmbed)
 
-# TODO: Error on missing perms (Manage Webhooks and Manage Messages) here
 @bot.command(aliases=["unsub"])
 @commands.check(subPerms)
 async def unsubscribe(ctx):
@@ -674,7 +230,7 @@ async def unsubscribe(ctx):
     unsubmsg = await ctx.send("Loading subscription list...")
 
     if "subDefault" not in servers[str(ctx.guild.id)][str(ctx.channel.id)]:
-        uInput = await subCheck(ctx, unsubmsg, 2, "Unsubscribe")
+        uInput = await subCheck(ctx, bot, unsubmsg, 2, "Unsubscribe")
     else:
         uInput = {
             "success": True,
@@ -757,7 +313,7 @@ async def unsubscribe(ctx):
                 if msg.content in ['1', '2', '3', '4', '5', '6', '7', '8', '9']:
                     with open("data/servers.json") as f:
                         servers = json.load(f)
-                    await getwebhook(servers, ctx.guild, ctx.channel)
+                    await getwebhook(bot, servers, ctx.guild, ctx.channel)
                     with open("data/servers.json") as f:
                         servers = json.load(f)
                     await msg.delete()
@@ -778,7 +334,7 @@ async def unsubscribe(ctx):
                 elif msg.content.lower() == 'a':
                     with open("data/servers.json") as f:
                         servers = json.load(f)
-                    await getwebhook(servers, ctx.guild, ctx.channel)
+                    await getwebhook(bot, servers, ctx.guild, ctx.channel)
                     with open("data/servers.json") as f:
                         servers = json.load(f)
                     for subType in uInput["subType"]:
@@ -813,7 +369,6 @@ async def unsub_error(ctx, error):
     if errEmbed:
         await ctx.send(embed=errEmbed)
 
-# TODO: Error on missing perms (Manage Messages) here
 @bot.command(aliases=["subs", "subslist", "subscriptions", "subscribed"])
 @commands.check(subPerms)
 async def sublist(ctx):
@@ -825,7 +380,7 @@ async def sublist(ctx):
     subsmsg = await ctx.send("Loading subscription list...")
 
     if "subDefault" not in servers[str(ctx.guild.id)][str(ctx.channel.id)]:
-        uInput = await getSubType(ctx, 2, subsmsg)
+        uInput = await getSubType(ctx, 2, bot, subsmsg)
     else:
         if not len(servers[str(ctx.guild.id)][str(ctx.channel.id)]["subDefault"]) > 1:
             uInput = {
@@ -960,7 +515,7 @@ async def mscheck(ctx, vtuber):
             "msText": subtext
         }
     }
-    await milestoneNotify(msDict)
+    await milestoneNotify(msDict, bot)
     await ctx.send(file=discord.File(f'milestone/generated/{vtuber}.png'))
 
 @bot.command()
@@ -970,7 +525,7 @@ async def nstest(ctx):
         servers = json.load(f)
     
     live = await streamcheck()
-    whurl = await getwebhook(servers, ctx.guild, ctx.channel)
+    whurl = await getwebhook(bot, servers, ctx.guild, ctx.channel)
 
     for livech in live:
         async with aiohttp.ClientSession() as session:
@@ -987,7 +542,7 @@ async def postas(ctx, vtuber, *, text):
         channels = json.load(f)
     with open("data/servers.json") as f:
         servers = json.load(f)
-    whurl = await getwebhook(servers, ctx.guild, ctx.channel)
+    whurl = await getwebhook(bot, servers, ctx.guild, ctx.channel)
     ytch = await channelInfo(channels[vtuber]["channel"])
     async with aiohttp.ClientSession() as session:
         webhook = Webhook.from_url(whurl, adapter=AsyncWebhookAdapter(session))
